Discussion functions

A group of functions which deal with discussion topics.

get_discussion_id: Get single discussion by id

update_discussion_id: Update discussion by id

get_discussions_context(object_id, object_type = "courses", include = NULL)

get_discussion_id(discussion_id, object_id, object_type = "courses")

update_discussion_id(discussion_id, object_id, message,
  object_type = "courses")

Arguments

object_id
course or group id

object_type
"courses" or "groups"

include
If "all_dates" is passed, all dates associated with graded discussions' assignments will be included.

discussion_id
specific id of discussion to get/update

message
new body of discussion id

Value

discussions belonging to requested context

single discussion

silently sends put request and updates

Details

get_discussions_context: get all discussions belonging to a course or group. Note that theoretically this should include announcements, as they are technically discussions, but does not. Use get_announcements instead.

Examples

get_discussions_context(4371405)

#> Error in canvas_query(url, args, "GET"): Unauthorized (HTTP 401).

get_discussion_id(4371405, 1350207)

#> Error in canvas_query(url, args, "GET"): Unauthorized (HTTP 401).

update_discussion_id(4371405, 1350207, newtext)

#> Error in update_discussion_id(4371405, 1350207, newtext): object 'newtext' not found